out_zerobus: use bundled FFI header
